Initial commit
Set up the start of a .vue checkout component that can be used to add
items to a shopping list and calculate rule based pricing. Also includes
base coverage testting to ensure that if any core DOM elements are
modified or the data structure changed or if any method names a re
removed or modified that the test will fail. For full test coverage I
would need to check that the functions within the buy(), remove() and
updateShoppingTotal() are tested for breaking changes.